// server/src/config.d.ts
export declare const CONFIG: {
    PORT: string | number;
    CHAIN_ID: number;
    RPC_URL: string;
    ENTRY_POINT_ADDRESS: string;
    RADRS_TOKEN_ADDRESS: string;
    PAYMASTER_ADDRESS: string;
    RADRS_FEE_RECEIVER: string;
    PAYMASTER_SIGNER_KEY: `0x${string}`;
    PRICE_RADRS_BNB: number;
    MARKUP: number;
    VALID_DURATION: number;
};
//# sourceMappingURL=config.d.ts.map